+++ enterprise/internal/batches/store/site_credentials.go
@@ -8,15 +8,21 @@ import (
 	btypes "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/types"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
 )
 
-func (s *Store) CreateSiteCredential(ctx context.Context, c *btypes.SiteCredential) error {
+func (s *Store) CreateSiteCredential(ctx context.Context, c *btypes.SiteCredential, credential auth.Authenticator) error {
 	if c.CreatedAt.IsZero() {
 		c.CreatedAt = s.now()
 	}
 
 	if c.UpdatedAt.IsZero() {
 		c.UpdatedAt = c.CreatedAt
+	}
+
+	c.Key = s.key
+	if err := c.SetAuthenticator(ctx, credential); err != nil {
+		return err
 	}
 
 	q := createSiteCredentialQuery(c)
@@ -28,19 +34,21 @@ func (s *Store) CreateSiteCredential(ctx context.Context, c *btypes.SiteCredenti
 var createSiteCredentialQueryFmtstr = `
 -- source: enterprise/internal/batches/store/site_credentials.go:CreateSiteCredential
 INSERT INTO
-	batch_changes_site_credentials (external_service_type, external_service_id, credential, created_at, updated_at)
+	batch_changes_site_credentials (external_service_type, external_service_id, credential, credential_enc, created_at, updated_at)
 VALUES
-	(%s, %s, %s, %s, %s)
+	(%s, %s, %s, %s, %s, %s)
 RETURNING
 	%s
 `
 
 func createSiteCredentialQuery(c *btypes.SiteCredential) *sqlf.Query {
+	unencrypted, encrypted := c.GetRawCredential()
 	return sqlf.Sprintf(
 		createSiteCredentialQueryFmtstr,
 		c.ExternalServiceType,
 		c.ExternalServiceID,
-		&database.NullAuthenticator{A: &c.Credential},
+		&database.NullAuthenticator{A: &unencrypted},
+		encrypted,
 		c.CreatedAt,
 		c.UpdatedAt,
 		sqlf.Join(siteCredentialColumns, ","),
@@ -86,7 +94,7 @@ type GetSiteCredentialOpts struct {
 func (s *Store) GetSiteCredential(ctx context.Context, opts GetSiteCredentialOpts) (*btypes.SiteCredential, error) {
 	q := getSiteCredentialQuery(opts)
 
-	var cred btypes.SiteCredential
+	cred := btypes.SiteCredential{Key: s.key}
 	err := s.query(ctx, q, func(sc scanner) error { return scanSiteCredential(&cred, sc) })
 	if err != nil {
 		return nil, err
@@ -129,6 +137,10 @@ func getSiteCredentialQuery(opts GetSiteCredentialOpts) *sqlf.Query {
 
 type ListSiteCredentialsOpts struct {
 	LimitOpts
+
+	// TODO(batch-changes-site-credential-encryption): remove when no longer
+	// needed.
+	OnlyUnencrypted bool
 }
 
 func (s *Store) ListSiteCredentials(ctx context.Context, opts ListSiteCredentialsOpts) (cs []*btypes.SiteCredential, next int64, err error) {
@@ -136,7 +148,7 @@ func (s *Store) ListSiteCredentials(ctx context.Context, opts ListSiteCredential
 
 	cs = make([]*btypes.SiteCredential, 0, opts.DBLimit())
 	err = s.query(ctx, q, func(sc scanner) (err error) {
-		var c btypes.SiteCredential
+		c := btypes.SiteCredential{Key: s.key}
 		if err := scanSiteCredential(&c, sc); err != nil {
 			return err
 		}
@@ -157,12 +169,69 @@ var listSiteCredentialsQueryFmtstr = `
 SELECT
 	%s
 FROM batch_changes_site_credentials
+WHERE %s
 ORDER BY external_service_type ASC, external_service_id ASC
 `
 
 func listSiteCredentialsQuery(opts ListSiteCredentialsOpts) *sqlf.Query {
+	preds := []*sqlf.Query{sqlf.Sprintf("TRUE")}
+	if opts.OnlyUnencrypted {
+		preds = append(preds, sqlf.Sprintf("credential_enc IS NULL"))
+	}
+
 	return sqlf.Sprintf(
 		listSiteCredentialsQueryFmtstr+opts.ToDB(),
+		sqlf.Join(siteCredentialColumns, ","),
+		sqlf.Join(preds, "AND"),
+	)
+}
+
+func (s *Store) UpdateSiteCredential(ctx context.Context, c *btypes.SiteCredential) error {
+	c.UpdatedAt = s.now()
+
+	updated := &btypes.SiteCredential{Key: s.key}
+	q := s.updateSiteCredentialQuery(c)
+	if err := s.query(ctx, q, func(sc scanner) error {
+		return scanSiteCredential(updated, sc)
+	}); err != nil {
+		return err
+	}
+
+	if updated.ID == 0 {
+		return ErrNoResults
+	}
+	*c = *updated
+	return nil
+}
+
+const updateSiteCredentialQueryFmtstr = `
+-- source: enterprise/internal/batches/store/site_credentials.go:UpdateSiteCredential
+UPDATE
+	batch_changes_site_credentials
+SET
+	external_service_type = %s,
+	external_service_id = %s,
+	credential = %s,
+	credential_enc = %s,
+	created_at = %s,
+	updated_at = %s
+WHERE
+	id = %s
+RETURNING
+	%s
+`
+
+func (s *Store) updateSiteCredentialQuery(c *btypes.SiteCredential) *sqlf.Query {
+	unencrypted, encrypted := c.GetRawCredential()
+	return sqlf.Sprintf(
+		updateSiteCredentialQueryFmtstr,
+		c.ExternalServiceType,
+		c.ExternalServiceID,
+		&database.NullAuthenticator{A: &unencrypted},
+		encrypted,
+		c.CreatedAt,
+		c.UpdatedAt,
+		c.ID,
 		sqlf.Join(siteCredentialColumns, ","),
 	)
 }
@@ -172,17 +241,30 @@ var siteCredentialColumns = []*sqlf.Query{
 	sqlf.Sprintf("external_service_type"),
 	sqlf.Sprintf("external_service_id"),
 	sqlf.Sprintf("credential"),
+	sqlf.Sprintf("credential_enc"),
 	sqlf.Sprintf("created_at"),
 	sqlf.Sprintf("updated_at"),
 }
 
 func scanSiteCredential(c *btypes.SiteCredential, sc scanner) error {
-	return sc.Scan(
+	var (
+		encrypted   []byte
+		unencrypted auth.Authenticator
+	)
+	na := database.NullAuthenticator{A: &unencrypted}
+
+	if err := sc.Scan(
 		&c.ID,
 		&c.ExternalServiceType,
 		&c.ExternalServiceID,
-		&database.NullAuthenticator{A: &c.Credential},
+		&na,
+		&encrypted,
 		&dbutil.NullTime{Time: &c.CreatedAt},
 		&dbutil.NullTime{Time: &c.UpdatedAt},
-	)
+	); err != nil {
+		return err
+	}
+
+	c.SetRawCredential(unencrypted, encrypted)
+	return nil
 }